Refactor Steam controllers and services, add inventory endpoints
Moved Steam Market item and search endpoints from SteamController to a new SteamMarketController. Introduced SteamInventoryController and SteamInventoryService for inventory-related operations, and removed the obsolete SteamProfileService. Updated ConfigHelper and SocialServiceHelper to use configurable Steam API and community URLs. Adjusted API routes to use the new controllers and endpoints.
